+++ src/scripts/core/config.js
@@ -59,14 +59,4 @@ export const BRTCONFIG = {
   },
 
   // END WORLD SETTINGS
-
-  // START DEFAULT SYSTEM SETTINGS
-
-  // QUANTITY_PROPERTY_PATH: "system.quantity",
-  // PRICE_PROPERTY_PATH: "system.price",
-  // SPELL_LEVEL_PATH: "system.level",
-  // ITEM_LOOT_TYPE: "loot",
-  // SCROLL_REGEX: /\s*Spell\s*Scroll\s*(\d+|cantrip)/i,
-
-  // END DEFAULT SYSTEM SETTINGS
 };