+++ src/scripts/harvest/harvest-chat-card.js
@@ -63,9 +63,11 @@ export class HarvestChatCard {
   async addToItemData(itemEntity, data) {
     const existingItem = this.itemsData.find((i) => i.item.id === itemEntity.id);
     const quantity = getProperty(data, BRTCONFIG.QUANTITY_PROPERTY_PATH) || 1;
+    const weight = getProperty(data, BRTCONFIG.WEIGHT_PROPERTY_PATH) || 0;
 
     if (existingItem) {
       existingItem.quantity = +existingItem.quantity + +quantity;
+      existingItem.weight = +existingItem.weight + +weight;
     } else {
       // we will scale down the font size if an item name is too long
       const fontSize = Math.max(60, 100 - Math.max(0, (itemEntity.name || itemEntity.text).length - 27) * 2);
@@ -94,6 +96,7 @@ export class HarvestChatCard {
           uuid: resultDoc?.uuid ?? "",
         },
         quantity: quantity,
+        weight: weight,
         fontSize: fontSize,
       });
     }