+++ src/scripts/systems/pf2e/pf2e.js
@@ -3,6 +3,7 @@ export default {
   DEFAULT_LOOT_SHEET: "pf2e.LootSheetPF2e",
   DEFAULT_SPELL_COMPENDIUM: "pf2e.spells-srd",
   QUANTITY_PROPERTY_PATH: "system.quantity.value",
+  WEIGHT_PROPERTY_PATH: "system.weight.value",
   PRICE_PROPERTY_PATH: "system.price.value",
   SPELL_LEVEL_PATH: "system.level.value",
   ITEM_LOOT_TYPE: "treasure",