+++ src/scripts/systems/sw5e/sw5e.js
@@ -3,6 +3,7 @@ export default {
   DEFAULT_LOOT_SHEET: "dnd5e.LootSheet5eNPC",
   DEFAULT_SPELL_COMPENDIUM: "dnd5e.spells",
   QUANTITY_PROPERTY_PATH: "system.quantity",
+  WEIGHT_PROPERTY_PATH: "system.weight",
   PRICE_PROPERTY_PATH: "system.price",
   SPELL_LEVEL_PATH: "system.level",
   ITEM_LOOT_TYPE: "loot",